Added 404 msgs

+++ src/controllers/contact.js
@@ -14,6 +14,7 @@ class ContactController {
   static async getAllContacts(req, res) {
     try {
       const contacts = await ContactService.getAllContacts();
+      if (contacts.length === 0) return out(res, 404, 'No Contacts found', null, 'NOT_FOUND');
       return out(res, 200, 'All Contacts', contacts);
     } catch (error) {
       return out(res, 500, error.message || error, null, 'SERVER_ERROR');
@@ -24,6 +25,7 @@ class ContactController {
     try {
       const { id } = req.params;
       const contact = await ContactService.getSingleContact({ _id: id });
+      if (!contact) return out(res, 404, 'No Contact found', null, 'NOT_FOUND');
       return out(res, 200, 'Contacts', contact);
     } catch (error) {
       return out(res, 500, error.message || error, null, 'SERVER_ERROR');
@@ -33,7 +35,8 @@ class ContactController {
   static async updateContact(req, res) {
     try {
       const { id } = req.params;
-      //   const contact = await ContactService.getSingleContact({ _id: id });
+      const contact = await ContactService.getSingleContact({ _id: id });
+      if (!contact) return out(res, 404, 'No Contact found', null, 'NOT_FOUND');
       const updatedContact = await ContactService.updateContact({ _id: id }, req.body);
       return out(res, 200, 'Contact updated', updatedContact);
     } catch (error) {
@@ -45,6 +48,7 @@ class ContactController {
     try {
       const { id } = req.params;
       const contact = await ContactService.getSingleContact({ _id: id });
+      if (!contact) return out(res, 404, 'No Contact found', null, 'NOT_FOUND');
       await ContactService.deleteContact(contact);
       return out(res, 200, 'Contact Deleted');
     } catch (error) {
